+++ Source/Arpeggiator.h
@@ -53,7 +53,7 @@ public:
 
    //IClickable
    void MouseReleased() override;
-   bool MouseMoved(float x, float y) override;
+   bool MouseMoved(double x, double y) override;
 
    //INoteReceiver
    void PlayNote(NoteMessage note) override;
@@ -83,12 +83,12 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(float& width, float& height) override
+   void GetModuleDimensions(double& width, double& height) override
    {
       width = mWidth;
       height = mHeight;
    }
-   void OnClicked(float x, float y, bool right) override;
+   void OnClicked(double x, double y, bool right) override;
 
    std::string GetArpNoteDisplay(int pitch);
    void UpdateInterval();
@@ -108,8 +108,8 @@ private:
    };
    std::vector<ArpNote> mChord;
 
-   float mWidth{ 100 };
-   float mHeight{ 20 };
+   double mWidth{ 100 };
+   double mHeight{ 20 };
 
    NoteInterval mInterval{ NoteInterval::kInterval_16n };
    int mLastPitch{ -1 };
