+++ Source/AudioSplitter.h
@@ -57,7 +57,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(float& w, float& h) override;
+   void GetModuleDimensions(double& w, double& h) override;
 
    std::vector<PatchCableSource*> mDestinationCables;
 };
