+++ Source/AudioToPulse.h
@@ -61,14 +61,14 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(float& width, float& height) override
+   void GetModuleDimensions(double& width, double& height) override
    {
       width = mWidth;
       height = mHeight;
    }
 
-   float mWidth{ 200 };
-   float mHeight{ 20 };
+   double mWidth{ 200 };
+   double mHeight{ 20 };
 
    FloatSlider* mThresholdSlider{ nullptr };
    FloatSlider* mReleaseSlider{ nullptr };
