+++ Source/BitcrushEffect.h
@@ -54,7 +54,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(float& width, float& height) override
+   void GetModuleDimensions(double& width, double& height) override
    {
       width = mWidth;
       height = mHeight;
@@ -67,6 +67,6 @@ private:
    FloatSlider* mCrushSlider{ nullptr };
    FloatSlider* mDownsampleSlider{ nullptr };
 
-   float mWidth{ 200 };
-   float mHeight{ 20 };
+   double mWidth{ 200 };
+   double mHeight{ 20 };
 };
