+++ Source/BufferShuffler.cpp
@@ -291,7 +291,7 @@ int BufferShuffler::GetNumSlices()
    return TheTransport->CountInStandardMeasure(mInterval) * TheTransport->GetTimeSigTop() / TheTransport->GetTimeSigBottom() * mNumBars;
 }
 
-void BufferShuffler::OnClicked(float x, float y, bool right)
+void BufferShuffler::OnClicked(double x, double y, bool right)
 {
    IDrawableModule::OnClicked(x, y, right);
