+++ Source/CanvasControls.h
@@ -70,9 +70,9 @@ private:
    //IDrawableModule
    void PreDrawModule() override;
    void DrawModule() override;
-   void GetModuleDimensions(float& width, float& height) override;
+   void GetModuleDimensions(double& width, double& height) override;
 
-   float mWidth{ 200 };
+   double mWidth{ 200 };
    Canvas* mCanvas{ nullptr };
    CanvasElement* mSelectedElement{ nullptr };
    ClickButton* mRemoveElementButton{ nullptr };
