+++ Source/ChordDisplayer.h
@@ -53,7 +53,7 @@ public:
 private:
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(float& width, float& height) override;
+   void GetModuleDimensions(double& width, double& height) override;
    bool mAdvancedDetection{ false };
    bool mUseScaleDegrees{ false };
    bool mShowIntervals{ false };
