+++ Source/ClipLauncher.h
@@ -76,7 +76,7 @@ private:
 
    //IDrawableModule
    void DrawModule() override;
-   void GetModuleDimensions(float& width, float& height) override;
+   void GetModuleDimensions(double& width, double& height) override;
 
    class SampleData
    {
